Remove "talk_config" table based configuration
We're using environment variables for now. These are easy and secure to
configure per site and can be used to store some of the settings of the
lets-talk app for now.

Another configuration approach has been discussed (see 3cb470e), but
that requires ability to read fastn variables somehow in .wasm modules.
This feature does not exist yet. Either fastn (host) or some package in
fifthtry-community should exist to stream line configuration.