Feat: WIP port template/ to template-final/
The template-final/ package name is renamed to
lets-talk-template.fifthtry.site as it's made to replace the old one. CI
will not push this directory yet, we'll update the CI to use this dir
(or remove template/ in favor of this) when this is ready.

This commit starts the port of the template/ (old lets-talk-template).
We use auth from the lets-auth.fifthtry.site so we don't need any auth
related files. Some things (specifically the api call to talk.wasm) does
not work yet.