Temporarly paus pushing app to fifthtry
Upcomings changes to app (lets-talk.fifthtry.site) might make it
incompatible with the current version of template/. It's better if we
pause the CI for sometime until we've updated the app and template both
and tested everything. Till then, people can use the old published code.

This is for a very short time.