Configure and use lets-talk in template-final/
A lot of the things are untested (storybook for example). The base
lets-talk configuration via lets-talk.ftd in template-final/ is working.

Everything in lets-talk.ftd is re-exporting things. This file exists to
allow template users to easily find the place for configuration.

Sessions data is still not moved to the app package yet.