(untested) meeting page
The contents of this file is basically copy pasted from the old
template/. It should not work most probably.

There are other things stopping use from reaching this page, the
lets-talk.fifthtry.site/actions/create-meeting#new-instant is not
sending meeting-page-url for some reason. This needs some debugging
before we can reach the meeting.ftd page and test it.