Feat: Make the meeting.ftd page work
The meeting page in the new template (and lets-talk.fifthtry.site
package) is now able to load meetings.

The session token is loaded via a client side fetch request. This is
stil good as we don't rely on a hardcoded mountpoint, we've access to
the `ftd.app_path_ex` js function in web component code.