Feat: Globally unique dyte participant ids
We used username of an account previously. This is unique on every
website but this does not address the situation where two different
website have user account with the same username. In this situation, the
meeting sessions of these two different accounts would be shared (both
will see each other's past sessions)

To fix this, we prefix the particpant id with the host name of the
website.

This is temporary fix until we've ability to serve this data through our
database.