Remove dependencies from old template/
The old template only used the build.js file from the lets-talk package.
The recent changes made in the package broke existing websites in
production because of this.

Preserving the old template with its build files is not needed now since
prod websites have been manually fixed but it's still usefull if someone
wants to try the old version of template.