Fix: ft_sdk::Scheme is buggy. Provide our impl
HTTPSScheme tries to be https always, except on 127.0.0.1.

The ft_sdk::Scheme relies on request uri whose scheme begins with
"wasm+proxy://" and not http:// or https:// so it can't determine and
stays at default (always http).

This is a problem in production where http request is sent under https
context. Browsers block this behaviour and it breaks the application.