Feat: Default to host if allowed_emails is empty
We match user's email against the host (website domain) if the
`who-can-create-meetings` variable is empty.

The error messages have also been improved:
- The error message will give you a link that you can use to receive a
  confirmation email to verify your account. This happens if your email
  matches the configuration of `who-can-create-meetings` but your
  account is not verified.

A bunch of tests have also been added to gain some confidence in logic.

For other cases, I'll later add an endpoint to check permissions and
hide the form UI based on it's value.