Feat: Extract some meeting data out of js into ftd
- Participant list
- Self (Me)
- Inside meeting or in the waiting room

The above information is now available in ftd (in state.ftd file). This
can be used to render the UI, we don't need Dyte's ui-kit components for
this.

The ftd variables are updated based on events received from Dyte
(meeting joined, left, etc).

The video/audio stream cannot be represented in ftd, the js will
directly manipulate `ftd.video` components to render video feeds in the
future (not part of this commit).