WIP Meeting grid with participant tiles
This rough code gives some ideas on how to build the layout. We will
have to pre determine each row height and width of each participant tile
based on the number of participants:

- 1:1 will make participant take full area and self will appear in
  bottom right
- 2x2 configuration for upto 4 people meeting
- 3x3 configuration for upto 9 people meeting
- 4x4 configuration for upto 16 people meeting

This does not account for situations like:
- when I pin some participant
- when someone is sharing their screen

On mobile phone screens, we can render a single column where the tile
size will shrink as the number of participants increase.
Screen shares should at least have a toggle full screen button so that
the meeting is usable on mobile devices.

This will take time, this commit just starts the ground work, that too
in not a good way.