Feat: `video-tile` UI component
This simple component creates a `ftd.video` that is used to show
participant's video feed. This same component can also be used for
"self" (since type of participant and self is same).

This is useful for people who want to create their own meeting page and
want the video tile to be setup correctly (correct id and player
settings).