Split js bundle | Render video feeds |  More State
The js output is split into two files now:
- build.js with `talk-app` and `dyte-*` web components
- functions.js with utility functions and functions to `join-meeting`.
  Other meeting related functions will be added to this file. A custom
  meeting page can call these functions when a meeting is initialized on
  the page.

The `talk-app` component now updates video feed of participants. This is
done by querying all the HTMLVideoElement(s) and attaching associated
participant's video feed to them.

`meeting-title` (string) and `meeting-initialized` (boolean) are now
available in state.ftd. The `meeting-inialized` variable indicates if a
global meeting object (on js window) is available or not.