Feat: `leave-meeting` and `end-meeting` functions
FTD wrapper functions that can be called to leave meeting or end meeting
for all.

The participant list is cleared when you leave meeting room as, for you,
there are no participants to render.

The permissions is checked by dyte internally so `end-meeting` can only
be triggered by meeting host (The one who created the meeting).