Use real state file by default for `meeting-page`
While overriding `meeting-page` it's easy to forget passing the `state:`
parameter because it has a default value (dummy-state) so you'll not see
errors. But the meeting UI does not work properly with dummy-data. The
default value should be the real state.ftd file.

Now even if you don't pass a `state:` param to our `meeting-page` it'll
pick up the correct one by default.