WIP: Render a `lets-talk.screen-tile` assuming an `active-participant`
`1-1-meeting-page` renders a large `screen-tile` component if there's an
`active-participant` (one who's sharing their screen). How this variable
is set is not clear ye.

One possibility is, from js, we set the active-participant to the first
participant from the list who's sharing their screen on every change to
participant list. We later render a UI list of all participants that are
sharing their screen and the user can click on one of them to set the
`active-participant`. This has not happened yet!