Use `lets-talk.control-bar` in `1-1-meeting-page`
The custom control-bar is removed in this file in favor of the
`lets-talk.control-bar`. `lets-talk.control-bar` will later have a
boolean parameter which will show a button to toggle rendering of the
`context-box` component.

The switch between the two views (screen sharing view and 1:1 video
view) will happen through the button to enable screen share.