Feat: Update or Clear `state.active-participant`
The UI switches to different views:
- one big screen share and two tiny participants view if there's an
  active-participant (!NULL)
- one big participant and one tiny self if active-participant == NULL

This `active-participant` variable is updated via the js web component
(talk-app). If no one is sharing their screen in the meeting, it clears
the `active-participant` otherwise it sets it to whoever is sharing
their screen.

The UI can render a list of participants that are sharing their screen
and set this variable value based on `on-click` on items of that list to
support multiple users sharing screens at a time in a meeting.