Fix: Use `lets-talk.site-logo` instead of `ds.site-logo`
There seems to be a bug:
Overriding `ds.site-logo` in lets-talk-template does not let
lets-talk.fifthtry.site package pick it up in it's usage of `ds`. I
haven't created a minimum reproducer for this yet.

For an immediate fix, a `lets-talk.site-logo` is introduced which the
`lets-talk-template` can override and it'll be displayed on our custom
meeting header on the meeting page.